Renaming console logger file.
