+++ src/bin/download.ts
@@ -3,7 +3,7 @@ import DataDir from '../lib/data/dir';
 import { DataSet } from '../lib/data/set';
 import { downloadAllData } from '../lib/engine/download';
 import { Hubspot } from '../lib/hubspot';
-import { ConsoleLogger } from '../lib/log/logger';
+import { ConsoleLogger } from '../lib/log/console';
 
 const log = new ConsoleLogger();
 const dataDir = DataDir.root.subdir("in");
