+++ src/lib/engine/summary.ts
@@ -1,5 +1,5 @@
 import { Deal } from "../hubspot/model/deal";
-import { ConsoleLogger } from "../log/logger";
+import { ConsoleLogger } from "../log/console";
 import { Table } from "../log/table";
 import { formatMoney, formatNumber } from "../util/formatters";
 import { isPresent } from "../util/helpers";
